Revert "Merge branch 'pinpoint' of https://github.com/jdhs-ftc/sparkfun-otos-quickstart"
This reverts commit 5d507aff21dcad1d0371dee809cb91b8378da90e, reversing
changes made to a4ac660c6b4cd18659e8d7a6f22a24b4008ab2a2.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TwoDeadWheelLocalizer.java
@@ -36,7 +36,7 @@ public final class TwoDeadWheelLocalizer implements Localizer {
     public final Encoder par, perp;
     public final IMU imu;
 
-    private double lastParPos, lastPerpPos;
+    private int lastParPos, lastPerpPos;
     private Rotation2d lastHeading;
 
     private final double inPerTick;
@@ -112,8 +112,8 @@ public final class TwoDeadWheelLocalizer implements Localizer {
             return new PoseVelocity2d(new Vector2d(0.0, 0.0), 0.0);
         }
 
-        double parPosDelta = parPosVel.position - lastParPos;
-        double perpPosDelta = perpPosVel.position - lastPerpPos;
+        int parPosDelta = parPosVel.position - lastParPos;
+        int perpPosDelta = perpPosVel.position - lastPerpPos;
         double headingDelta = heading.minus(lastHeading);
 
         Twist2dDual<Time> twist = new Twist2dDual<>(
